Factor out serialization.ts.
